+++ Assets/YooAsset/Editor/AssetBundleDebugger/VisualViewers/DebuggerBundleListViewer.cs
@@ -147,7 +147,7 @@ namespace YooAsset.Editor
 
 			// Status
 			StyleColor textColor;
-			if (bundleInfo.Status == AssetBundleLoaderBase.EStatus.Failed)
+			if (bundleInfo.Status == (int)AssetBundleLoaderBase.EStatus.Failed)
 				textColor = new StyleColor(Color.yellow);
 			else
 				textColor = label1.style.color;
